Makes code compiles OK with Windows 7 SDK 7.1 (VC++10)

+++ include/libxls/ole.h
@@ -182,7 +182,7 @@ PSS;
 
 #pragma pack(pop)
 
-ssize_t ole2_read(void* buf,size_t size,size_t count,OLE2Stream* olest);
+int64_t ole2_read(void* buf,size_t size,size_t count,OLE2Stream* olest);
 OLE2Stream* ole2_sopen(OLE2* ole,DWORD start, size_t size);
 int ole2_seek(OLE2Stream* olest,DWORD ofs);
 OLE2Stream*  ole2_fopen(OLE2* ole, const char *file);
